requirements.txt
Flask==2.0.1
requests==2.25.1
python-pptx==0.6.21
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.29.0
gunicorn==20.1.0
